hello-next-js - chore: move global into  src/lib folder
