+++ myapps/hello-next-js/pages/api/tasks/v1/sql/create-row.ts
@@ -1,6 +1,6 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
 import { db } from '@/lib/db/db_postgreSQL';
-import { CHECK_BFF_AUTHORIZATION } from '../../../../../global/common';
+import { CHECK_BFF_AUTHORIZATION } from '@/global/common';
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
     await CHECK_BFF_AUTHORIZATION(req, res);
