+++ myapps/hello-next-js/pages/api/tasks/v1/sql/delete-rows.ts
@@ -9,7 +9,7 @@ Use DELETE when:
 
 import type { NextApiRequest, NextApiResponse } from 'next';
 import { db } from '@/lib/db/db_postgreSQL';
-import { CHECK_BFF_AUTHORIZATION } from '../../../../../global/common';
+import { CHECK_BFF_AUTHORIZATION } from '@/global/common';
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
     await CHECK_BFF_AUTHORIZATION(req, res);
