+++ myapps/hello-next-js/pages/api/tasks/v1/sql/graphql.ts
@@ -3,7 +3,7 @@ import { NextApiRequest, NextApiResponse } from 'next';
 import { db } from '@/lib/db/db_postgreSQL';
 import { Task } from "@/app/types/Task";
 import { values, placeholders } from "./seed-table";
-import { CHECK_BFF_AUTHORIZATION } from '../../../../../global/common';
+import { CHECK_BFF_AUTHORIZATION } from '@/global/common';
 
 import { InMemoryLRUCache } from '@apollo/utils.keyvaluecache';
