+++ myapps/hello-next-js/pages/api/tasks/v1/sql/index.ts
@@ -1,6 +1,6 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
 import { db } from '@/lib/db/db_postgreSQL';
-import { CHECK_BFF_AUTHORIZATION } from '../../../../../global/common';
+import { CHECK_BFF_AUTHORIZATION } from '@/global/common';
 
 /**
  * @swagger
