+++ myapps/hello-next-js/pagesLayouts/Footer.tsx
@@ -2,7 +2,7 @@ import { memo } from 'react';
 import Image from 'next/image';
 import Link from 'next/link';
 import styles from "@/app/page.module.css";
-import { MONOREPO_PREFIX } from '../global/common';
+import { MONOREPO_PREFIX } from '@/global/common';
 
 
 const Footer = () => {
