+++ myapps/hello-next-js/src/app/models/Task/use-client/TaskModel.test.ts
@@ -1,5 +1,5 @@
 import { TaskModel } from './TaskModel';
-import { BASE_URL } from "../../../../../global/common";
+import { BASE_URL } from "@/global/common";
 
 // Define mock response type
 type MockResponse = {
