+++ myapps/hello-next-js/src/app/views/Task/use-client/taskFilterWithDeferredValue.tsx
@@ -2,7 +2,7 @@
 import { useState, useMemo, useDeferredValue } from 'react';
 import { TaskTableType } from '@/app/types/Task';
 import { TaskTable } from '@/app/components/TaskTable';
-import { TASKS_CRUD } from '../../../../../global/common';
+import { TASKS_CRUD } from '@/global/common';
 import Link from "next/link";
 
 // dev note - use-case scenario of useDeferredValue:
