+++ myapps/hello-next-js/src/app/views/Task/use-client/taskWithSWRPage.tsx
@@ -1,10 +1,10 @@
 'use client';
 import { useState, useRef, useEffect } from "react";
-import { useTaskViewModelWithSwr } from '../../../viewModels/Task/use-client/useTasksViewModelWithSwr';
+import { useTaskViewModelWithSwr } from '@/app/viewModels/Task/use-client/useTasksViewModelWithSwr';
 import { TaskSeedDB } from '@/app/components/TaskSeedDB';
 import { TaskTable } from '@/app/components/TaskTable';
 import { Task } from "@/app/types/Task";
-import { TASKS_CRUD } from "../../../../../global/common";
+import { TASKS_CRUD } from "@/global/common";
 import Link from "next/link";
 
 export const TaskWithSWRPage = () => {
