+++ myapps/hello-next-js/src/global/awsParameterStore.ts
