+++ myapps/hello-next-js/src/global/awsSecretManager.ts
