+++ myapps/hello-next-js/src/global/common.ts
